rework SetpointSensorPair
