+++ src/plugins/spark/components/BlockWidgetDialogToolbar.vue
@@ -1,5 +1,5 @@
 <script lang="ts">
-import { Component, Emit, Prop } from 'vue-property-decorator';
+import { Component, Prop } from 'vue-property-decorator';
 
 import { WidgetMode } from '@/store/features';
