+++ src/plugins/spark/features/SetpointSensorPair/SetpointSensorPairWidget.vue
@@ -5,9 +5,16 @@ import { Component } from 'vue-property-decorator';
 import BlockWidgetBase from '@/plugins/spark/components/BlockWidgetBase';
 
 import { sparkStore } from '../../store';
+import SetpointSensorPairBasic from './SetpointSensorPairBasic.vue';
+import SetpointSensorPairFull from './SetpointSensorPairFull.vue';
 import { SetpointSensorPairBlock } from './types';
 
-@Component
+@Component({
+  components: {
+    Basic: SetpointSensorPairBasic,
+    Full: SetpointSensorPairFull,
+  },
+})
 export default class SetpointSensorPairWidget extends BlockWidgetBase {
   readonly block!: SetpointSensorPairBlock;
 
@@ -19,61 +26,36 @@ export default class SetpointSensorPairWidget extends BlockWidgetBase {
 </script>
 
 <template>
-  <q-card dark class="text-white scroll">
-    <BlockWidgetToolbar :crud="crud" />
+  <GraphCardWrapper :show="inDialog">
+    <template #graph>
+      <HistoryGraph :id="widget.id" :config="graphCfg" />
+    </template>
 
-    <CardWarning v-if="!isUsed">
-      <template #message>
-        This Setpoint is not used as PID input.
+    <component :is="mode" :crud="crud" :class="cardClass">
+      <template #toolbar>
+        <component :is="toolbarComponent" :crud="crud" :mode.sync="mode" />
       </template>
-    </CardWarning>
-    <CardWarning v-else-if="!block.data.settingEnabled">
-      <template #message>
-        <span>This setpoint is disabled.</span>
-      </template>
-      <template #actions>
-        <q-btn
-          text-color="white"
-          flat
-          label="Enable"
-          @click="block.data.settingEnabled = true; saveBlock();"
-        />
-      </template>
-    </CardWarning>
 
-    <q-card-section>
-      <q-item dark>
-        <q-item-section class="q-mr-md">
-          <q-item-label caption>
-            Setting
-          </q-item-label>
-          <UnitField
-            :class="{darkened: !block.data.settingEnabled}"
-            :value="block.data.storedSetting"
-            :readonly="isDriven"
-            title="Setting"
-            tag="big"
-            @input="v => {block.data.storedSetting = v; saveBlock()}"
-          />
-        </q-item-section>
-        <q-item-section class="q-mr-md">
-          <q-item-label caption>
-            Sensor
-          </q-item-label>
-          <UnitField :value="block.data.value" tag="big" readonly />
-        </q-item-section>
-        <q-item-section class="col-auto">
-          <q-item-label caption>
-            Unfiltered sensor
-          </q-item-label>
-          <UnitField :value="block.data.valueUnfiltered" tag="big" readonly />
-        </q-item-section>
-      </q-item>
-      <q-item dark>
-        <q-item-section>
-          <DrivenIndicator :block-id="block.id" :service-id="serviceId" />
-        </q-item-section>
-      </q-item>
-    </q-card-section>
-  </q-card>
+      <template #warnings>
+        <CardWarning v-if="!isUsed">
+          <template #message>
+            This Setpoint is not used as PID input.
+          </template>
+        </CardWarning>
+        <CardWarning v-else-if="!block.data.settingEnabled">
+          <template #message>
+            <span>This setpoint is disabled.</span>
+          </template>
+          <template #actions>
+            <q-btn
+              text-color="white"
+              flat
+              label="Enable"
+              @click="block.data.settingEnabled = true; saveBlock();"
+            />
+          </template>
+        </CardWarning>
+      </template>
+    </component>
+  </GraphCardWrapper>
 </template>
