+++ src/plugins/spark/features/SetpointSensorPair/index.ts
@@ -1,10 +1,10 @@
 import { Link, Unit } from '@/helpers/units';
+import { interfaceTypes } from '@/plugins/spark/block-types';
 import GenericBlock from '@/plugins/spark/components/GenericBlock';
+import { blockWidgetSelector } from '@/plugins/spark/helpers';
+import { BlockSpec } from '@/plugins/spark/types';
 import { Feature } from '@/store/features';
 
-import { interfaceTypes } from '../../block-types';
-import { blockWidgetSelector } from '../../helpers';
-import { BlockSpec } from '../../types';
 import { typeName } from './getters';
 import widget from './SetpointSensorPairWidget.vue';
 import { FilterChoice, SetpointSensorPairData } from './types';
